led_strip_monitor_pokus - Copy/requirements.txt
pyaudiowpatch
PyQt6
pyserial
Pillow
numpy
mss
dxcam
keyboard
psutil
opencv-python
netifaces
zeroconf
sounddevice
